aicardio/requirements.txt
albumentations==1.1.0
easydict==1.9
Flask==2.1.0
loguru==0.5.3
matplotlib==3.3.4
numpy==1.19.5
opencv_python_headless==4.5.5.62
Pillow==9.0.0
pydicom==2.2.2
pykalman==0.9.5
PyYAML==6.0
redis==4.1.1
segmentation==0.2.2
terminaltables==3.1.10
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.62.3
Werkzeug==0.16.1
